+++ relatives/tests/admin.py
@@ -26,3 +26,6 @@ admin.site.register(models.Pet, PetAdmin)
 
 admin.site.register(models.Movie)
 admin.site.register(models.Actor)
+admin.site.register(models.Book)
+admin.site.register(models.Image)
+admin.site.register(models.Journal)